remove every warnings

+++ main.cpp
@@ -2,9 +2,9 @@
 
 void printDoubleVector(vector<vector<string>> test)
 {
-    for(int i = 0; i < test.size(); i++)
+    for(size_t i = 0; i < test.size(); i++)
     {
-        for (int j = 0; j < test[i].size(); j++)
+        for (size_t j = 0; j < test[i].size(); j++)
         {
             cout << test[i][j];
             cout << " ";
